+++ src/Common/Constants.php
@@ -9,4 +9,9 @@ class Constants
     public const MIN_LIMIT = 10;
     public const MAX_LIMIT = 100;
     public const POSSIBLE_EVENT_TYPES = ['processed', 'queued', 'sent', 'delivered', 'soft_bounced', 'hard_bounced', 'junk', 'opened', 'clicked', 'unsubscribed', 'spam_complaints'];
+    public const POSSIBLE_GROUP_BY_OPTIONS = ['days', 'weeks', 'months', 'years'];
+    public const GROUP_BY_DAYS = 'days';
+    public const GROUP_BY_WEEKS = 'weeks';
+    public const GROUP_BY_MONTHS = 'months';
+    public const GROUP_BY_YEARS = 'years';
 }